move helper methods to module, assign methods to variables
