+++ lib/concerns/helper.rb
@@ -1,13 +1,39 @@
 module Helper
-  def format_song_params(filename)
-    split_name = split_filename filename
-    artist=(Artist.find_or_create_by_name split_name[1])
-    genre=(Genre.find_or_create_by_name split_name[2])
-    [split_name[0], artist, genre]
+  def get_songs_by_type(type, name)
+    model = type.find_by_name(name)
+    song_list = model.songs.map { |song| format_song song } unless model.nil?
+    show_results song_list || []
   end
 
-  def split_filename(filename)
-    split_name = filename.gsub('.mp3', '').split(" - ")
-    [split_name[1], split_name[0], split_name[2]]
+  def show_results(result_list)
+    if result_list.empty?
+      puts "No results found."
+      return continue_message
+    end
+    format_message result_list
+    continue_message
+  end
+
+  def continue_message
+    puts "\nWhat else would you like me to do for you? "\
+      "Type 'help' to view commands"
+  end
+
+  def begin_message
+    if @command.nil?
+      message = "Hi #{ENV['USER']}, I am your music buddy."\
+        "What would you like me to do for you?\n"\
+        "Type 'help' to view commands"
+      format_message message
+    end
+  end
+
+  def get_track
+    track = Integer(gets.chomp) rescue nil
+    while track.nil? || Song.all.size < track || track < 1
+      puts "Please enter a valid track number"
+      track = Integer(gets.chomp) rescue nil
+    end
+    format_song Song.all[track - 1]
   end
 end
